Split workspace dialog families
