+++ Sources/QuillCodeApp/QuillCodeSearchAndShortcutDialogs.swift
@@ -1,0 +1,179 @@
+import SwiftUI
+import QuillCodeCore
+
+struct QuillCodeKeyboardShortcutsView: View {
+    var commands: [WorkspaceCommandSurface]
+    var onClose: () -> Void
+
+    private var shortcutCommands: [WorkspaceCommandSurface] {
+        commands.filter { $0.shortcut?.isEmpty == false }
+    }
+
+    private var groups: [WorkspaceCommandGroupSurface] {
+        WorkspaceCommandPalette.groupedCommands(shortcutCommands, matching: "")
+    }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 16) {
+            QuillCodeDialogHeader(
+                title: "Keyboard shortcuts",
+                subtitle: "Fast paths for the workspace actions available right now.",
+                closeTitle: "Close",
+                onClose: onClose
+            )
+
+            ScrollView {
+                LazyVStack(alignment: .leading, spacing: 14) {
+                    ForEach(groups) { group in
+                        VStack(alignment: .leading, spacing: 6) {
+                            QuillCodeDialogSectionTitle(group.title)
+                            ForEach(group.commands) { command in
+                                QuillCodeShortcutRow(command: command)
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        .padding(20)
+        .frame(width: 560, height: 520)
+        .background(QuillCodePalette.background)
+    }
+}
+
+struct QuillCodeSearchView: View {
+    var sidebar: SidebarSurface
+    @Binding var query: String
+    var onSelectThread: (UUID) -> Void
+    var onClose: () -> Void
+
+    @FocusState private var isSearchFocused: Bool
+
+    private var results: [SidebarItemSurface] {
+        sidebar.filteredItems(matching: query)
+    }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 16) {
+            QuillCodeDialogHeader(
+                title: "Search chats",
+                subtitle: "Find a thread by title, model, pinned state, archived state, or transcript text.",
+                closeTitle: "Close",
+                onClose: onClose
+            )
+
+            TextField("Search chats", text: $query)
+                .textFieldStyle(.roundedBorder)
+                .focused($isSearchFocused)
+                .frame(minHeight: QuillCodeMetrics.minimumHitTarget)
+                .onSubmit {
+                    if let firstResult = results.first {
+                        onSelectThread(firstResult.id)
+                    }
+                }
+
+            if results.isEmpty {
+                QuillCodeDialogEmptyState(
+                    systemImage: "magnifyingglass",
+                    title: "No matching chats",
+                    subtitle: "Try a thread title, selected model, pinned, or prior message text."
+                )
+            } else {
+                ScrollView {
+                    LazyVStack(spacing: 8) {
+                        ForEach(results) { item in
+                            QuillCodeSearchResultRow(item: item, onSelect: onSelectThread)
+                        }
+                    }
+                }
+            }
+        }
+        .padding(20)
+        .frame(width: 560, height: 520)
+        .background(QuillCodePalette.background)
+        .onAppear {
+            DispatchQueue.main.async {
+                isSearchFocused = true
+            }
+        }
+        .onDisappear {
+            isSearchFocused = false
+        }
+    }
+}
+
+private struct QuillCodeShortcutRow: View {
+    var command: WorkspaceCommandSurface
+
+    var body: some View {
+        HStack(spacing: 12) {
+            VStack(alignment: .leading, spacing: 3) {
+                Text(command.title)
+                    .font(.callout.weight(.semibold))
+                    .foregroundStyle(command.isEnabled ? QuillCodePalette.text : QuillCodePalette.muted)
+                    .lineLimit(1)
+                if !command.keywords.isEmpty {
+                    Text(command.keywords.prefix(3).joined(separator: " - "))
+                        .font(.caption)
+                        .foregroundStyle(QuillCodePalette.muted)
+                        .lineLimit(1)
+                }
+            }
+            Spacer()
+            Text(command.shortcut ?? "")
+                .font(.caption.monospacedDigit().weight(.semibold))
+                .foregroundStyle(QuillCodePalette.text)
+                .padding(.horizontal, 10)
+                .padding(.vertical, 6)
+                .background(QuillCodePalette.selection)
+                .clipShape(RoundedRectangle(cornerRadius: 8, style: .continuous))
+        }
+        .padding(12)
+        .frame(minHeight: QuillCodeMetrics.minimumHitTarget)
+        .background(QuillCodePalette.panel)
+        .overlay(
+            RoundedRectangle(cornerRadius: 12, style: .continuous)
+                .stroke(Color.white.opacity(0.08))
+        )
+        .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
+    }
+}
+
+private struct QuillCodeSearchResultRow: View {
+    var item: SidebarItemSurface
+    var onSelect: (UUID) -> Void
+
+    var body: some View {
+        Button {
+            onSelect(item.id)
+        } label: {
+            HStack(spacing: 12) {
+                Image(systemName: item.isPinned ? "pin.fill" : "text.bubble")
+                    .foregroundStyle(item.isSelected ? QuillCodePalette.blue : QuillCodePalette.muted)
+                    .frame(width: 22)
+                    .accessibilityHidden(true)
+                VStack(alignment: .leading, spacing: 4) {
+                    Text(item.title)
+                        .font(.callout.weight(.semibold))
+                        .lineLimit(1)
+                    Text(item.subtitle + (item.isPinned ? " - pinned" : "") + (item.isArchived ? " - archived" : ""))
+                        .font(.caption)
+                        .foregroundStyle(QuillCodePalette.muted)
+                        .lineLimit(1)
+                }
+                Spacer()
+                if item.isSelected {
+                    Image(systemName: "checkmark")
+                        .foregroundStyle(QuillCodePalette.blue)
+                        .accessibilityHidden(true)
+                }
+            }
+            .padding(12)
+            .frame(maxWidth: .infinity, minHeight: QuillCodeMetrics.minimumHitTarget, alignment: .leading)
+            .background(item.isSelected ? QuillCodePalette.selection : QuillCodePalette.panel)
+            .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
+            .contentShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
+        }
+        .buttonStyle(QuillCodePressableButtonStyle())
+    }
+}
