+++ docs/DECISIONS.md
@@ -176,7 +176,7 @@
 - Terminal surface contracts live in `QuillCodeTerminalSurface.swift`. `TerminalSurface` and `TerminalCommandSurface` own run/clear availability, cwd labels, command lifecycle labels, and execution-context propagation; `WorkspaceSurface.swift` should only carry the aggregate `terminal` payload and assemble it from terminal state.
 - Secondary utility panes live in `QuillCodeSecondaryPanesView.swift`. Extensions, Memories, Automations, their count/status pills, empty states, and compact cards should evolve there; `WorkspaceSwiftUIView.swift` should only decide pane placement and route commands.
 - Secondary pane surface contracts live in `QuillCodeSecondaryPaneSurface.swift`. `WorkspaceExtensionsSurface`, `WorkspaceMemoriesSurface`, `WorkspaceAutomationsSurface`, extension manifest rows, memory rows, and automation workflow rows should evolve beside the secondary-pane native and HTML renderers; `WorkspaceSurface.swift` should only carry the aggregate slots and assemble them from model state.
-- Workspace modal surfaces live in `QuillCodeWorkspaceDialogs.swift`. Command palette, keyboard shortcuts, chat search, rename sheets, and worktree create/remove sheets should evolve there; `WorkspaceSwiftUIView.swift` should only decide when a sheet is presented and route completed actions.
+- Workspace modal presentation lives in `QuillCodeWorkspaceSheets.swift`, while each dialog family has a named owner: `QuillCodeCommandPaletteDialog.swift` owns command palette rows and icon mapping, `QuillCodeSearchAndShortcutDialogs.swift` owns chat search and keyboard shortcuts, `QuillCodeWorktreeDialogs.swift` owns worktree create/remove drafts and sheets, `QuillCodeWorkspaceDialogs.swift` owns rename sheets, and `QuillCodeDialogChrome.swift` owns shared dialog header, section, empty-state, and labeled-field primitives. `WorkspaceSwiftUIView.swift` should only decide when sheets are presented and route completed actions.
 - Workspace view command routing is planned by `WorkspaceViewCommandPlanner`. The planner owns command-ID to view-action mapping, selected thread/project rename lookup, worktree sheet intents, and composer focus rules. `WorkspaceSwiftUIView.swift` should execute typed `WorkspaceViewCommandAction` values and avoid growing another command-ID switch as Codex-parity commands expand.
 - Sidebar row actions are planned by `WorkspaceSidebarRowActionPlanner`. The planner owns thread/project rename lookups and maps non-rename row menu choices to typed `WorkspaceThreadRowMutation` and `WorkspaceProjectRowMutation` values. SwiftUI should only open rename sheets or forward typed mutations, while the desktop controller should execute those mutations through `WorkspaceSidebarRowMutationExecutor` instead of switching over row action enums.
 - Sidebar bulk actions are split into three layers: `WorkspaceSidebarSelectionEngine` owns selection state, `WorkspaceSidebarBulkActionPlanner` maps commands to mutation plans, and `WorkspaceSidebarBulkActionExecutor` applies those plans to thread/project selection values. `WorkspaceModel.swift` should keep persistence, terminal sync, and top-bar refresh side effects, but it should not inline pin/archive/delete mutation logic.
